TimePix MZI Data analysis pipline
This is all the code used to analyze all the spot interference information for the interferometer setup, you just need to change the initial variables, perhaps change the defined regions and associated wavelength values, and the rest should work off of that. Any questions contact me at eklein550@gmail.com!